refactor: split activity log repository
