+++ .agents/memory-bank/activeContext.md
@@ -6,6 +6,7 @@ The current branch is `codex/realm-convergence`. Keep this file focused on the a
 ## Current Branch Projects
 These records stay active until the branch is ready to merge. Before merge, move completed project records into the matching history file and leave only the next focus, constraints, and history index here.
 
+- **ActivityLogRepository File Boundary Split (2026-07-03, complete):** Split activity log repository declarations out of `AnnotationRepository.swift` into `Models/Repositories/ActivityLogRepository.swift`, leaving the annotation repository file focused on bookmarks/highlights/annotations. `ActivityLogUIEntry`, `ActivityLogRepositoryProtocol`, and `RealmActivityLogRepository` keep their existing external API and UI behavior. `RealmActivityLogRepository.mapToUI` no longer constructs `CalibreBookRealm.PrimaryKey`; it resolves the `CalibreLibrary` from a `CalibreLibrarySnapshotProviding` snapshot and uses the existing `BookRepositoryProtocol.getBook(library:bookId:)` value lookup to fill book titles. The existing `container:` initializer remains for AppContainer call sites, and a narrower `librarySnapshotProvider:` initializer supports lightweight repository tests without constructing the full app container. Xcode project target membership was updated for iOS and Catalyst. Validation: activity-log split scans passed with no activity declarations left in `AnnotationRepository.swift` and no `CalibreBookRealm.PrimaryKey` in `ActivityLogRepository.swift`; focused `ActivityLogRepositoryTests`, `CalibreActivityLoggerTests`, `ActivityListViewModelTests`, and `BookDetailViewModelTests` passed (30 tests); standard iOS Simulator build passed; `git diff --check` passed.
 - **CalibreServerDSReaderHelper Realm Boundary Shrink (2026-07-03, complete):** Converted outward-facing `CalibreServerDSReaderHelper` into a pure `Hashable` value type in `CalibrePluginModels.swift`, carrying `port`, `configurationData`, and `CalibreDSReaderHelperConfiguration` encode/decode behavior without Realm. Renamed the Realm embedded object implementation to `CalibreServerDSReaderHelperRealm` and kept it inside the Realm boundary with `value`/`apply`/`toValue` mapper methods; the Realm runtime class name remains `CalibreServerDSReaderHelper` to avoid a schema rename and no schema version bump was made. `ServerRepositoryProtocol` method names stayed unchanged while `RealmServerRepository` now maps between value and Realm object internally. View, Network, Manager, and AppContainer call sites continue to work with the pure value, and tests that write `CalibreServerRealm.dsreaderHelper` directly now do so through `CalibreServerDSReaderHelperRealm(value:)`. Validation: DSReader Helper Realm boundary scans passed with no app-layer `CalibreServerDSReaderHelperRealm` leakage and no new `RealmSwift` imports in Views/Network/Managers; focused `RealmServerRepositoryTests`, `RealmDomainMappingTests`, `CalibreServerManagerTests`, `DSReaderHelperConnectorTests`, and `BookDetailViewModelTests` passed (63 tests before the mapper guard, then `RealmServerRepositoryTests` passed with the added schema-class-name mapper test); standard iOS Simulator build passed; `git diff --check` passed.
 - **Repositories Realm Boundary Analysis (2026-07-03, findings recorded):** Direct Realm usage inside concrete `Realm*Repository` implementations is mostly expected, but the next unreasonable boundaries are now identified. Highest priority: `CalibreServerDSReaderHelper` is a Realm `EmbeddedObject` but is exposed through `ServerRepositoryProtocol` and flows into `CalibreServerManager`, `CalibreBookManager`, `ServerViewModel`, and `DSReaderHelperConnector`; it should become a pure value type at the repository boundary with Realm mapping kept under `Models/Realm`. Second priority: `RealmActivityLogRepository` and `ActivityLogRepositoryProtocol` currently live inside `AnnotationRepository.swift`; activity log should be split into its own repository file/boundary, and `ActivityLogUIEntry` is UI-shaped so longer-term repository output should become a pure activity-log value with view-model presentation mapping. Third priority: `RealmActivityLogRepository.mapToUI` still builds `CalibreBookRealm.PrimaryKey` directly despite depending on `BookRepositoryProtocol`; use pure repository lookup helpers instead. Fourth priority: `RealmReadingPositionRepository.historyBook(...)` crosses from reading-position sidecar storage into the main metadata Realm through `CalibreBookRealm`; move that lookup to `BookRepositoryProtocol` or have callers compose position history with book repository reads. Lower-priority cleanup: mapper/helper code such as `BookBookmarkRealm` / `BookHighlightRealm` convenience mapping in `AnnotationRepository.swift` and `BookDeviceReadingPositionRealm.identity` in `ReadingPositionRepository.swift` should move back to `Models/Realm/*Mappers.swift`. `RealmSearchCacheStore` and `ServerScopedRealmConfigurationProviding` are acceptable Realm boundaries for now, though `RealmSearchCacheStore` remains a directory/naming inconsistency because it owns Realm cache implementation while living under `Models/Repositories`.
 - **Managers Realm Boundary Shrink (2026-07-03, complete):** Removed direct Realm usage from `Models/Managers/`. `CalibreBookManager` no longer imports `RealmSwift`, opens Realm, references `CalibreBookRealm` / `CalibreLibraryRealm`, or exposes Realm bridge conversion methods; book primary-key creation, local-domain lookup, metadata entry persistence, sync-list writes, and deleted-book cleanup now go through `BookRepositoryProtocol` / `RealmBookRepository` with pure value inputs and `BookMetadataPersistenceResult` outputs. `ReadingSessionManager` now delegates default reading-history Realm queries to `ReadingPositionRepositoryProtocol.positionHistory(...)`. `CalibreLibraryManager` sync code now passes `BookMetadataSyncRecord` values instead of Realm dictionaries, and manager method names were cleaned from `*Realm` wrappers to repository-oriented save/delete names. The unused network-side `CalibreServerService.handleLibraryBookOne(library:bookRealm:...)` Realm bridge was removed. Validation: Managers Realm scans passed with no matches; focused `CalibreBookManagerTests`, `CalibreLibraryManagerTests`, `ReadingSessionManagerTests`, `RealmBookRepositoryTests`, `RealmReadingPositionRepositoryTests`, `BookDetailViewModelTests`, `UnifiedSearchServiceTests`, and `UnifiedSearchIntegrationTests` passed (91 tests); standard iOS Simulator build passed; `git diff --check` passed.
