+++ YetAnotherEBookReaderTests/ActivityLogRepositoryTests.swift
@@ -9,24 +9,30 @@ import RealmSwift
 
 final class ActivityLogRepositoryTests: XCTestCase {
     private var databaseService: DatabaseService!
+    private var bookRepository: MockBookRepository!
+    private var librarySnapshotProvider: MockCalibreLibrarySnapshotProvider!
     private var repository: RealmActivityLogRepository!
 
     override func setUpWithError() throws {
         try super.setUpWithError()
         databaseService = DatabaseService()
+        bookRepository = MockBookRepository()
+        librarySnapshotProvider = MockCalibreLibrarySnapshotProvider()
         let config = MockDatabaseService.inMemoryConfiguration(
             identifier: "ActivityLogRepositoryTests-\(UUID().uuidString)"
         )
         databaseService.installTestConfiguration(config)
         repository = RealmActivityLogRepository(
             databaseService: databaseService,
-            bookRepository: MockBookRepository(),
+            bookRepository: bookRepository,
             container: nil
         )
     }
 
     override func tearDownWithError() throws {
         repository = nil
+        librarySnapshotProvider = nil
+        bookRepository = nil
         databaseService = nil
         try super.tearDownWithError()
     }
@@ -70,6 +76,44 @@ final class ActivityLogRepositoryTests: XCTestCase {
         XCTAssertEqual(entries.first?.httpBodyString, "payload")
     }
 
+    func testFetchEntriesMapsBookTitleThroughBookRepositoryLibraryLookup() async throws {
+        let config = MockDatabaseService.inMemoryConfiguration(
+            identifier: "ActivityLogRepositoryMappingTests-\(UUID().uuidString)"
+        )
+        databaseService.installTestConfiguration(config)
+
+        let library = TestFixtures.makeLibrary()
+        var book = TestFixtures.makeBook(id: 42, library: library)
+        book.title = "Mapped Activity Book"
+        librarySnapshotProvider.calibreLibraries = [library.id: library]
+        bookRepository.getBookReturn = book
+        repository = RealmActivityLogRepository(
+            databaseService: databaseService,
+            bookRepository: bookRepository,
+            librarySnapshotProvider: librarySnapshotProvider
+        )
+
+        let request = URLRequest(url: URL(string: "http://calibre.local/ajax/books")!)
+        await repository.writeActivityLogEvents([
+            .start(
+                ActivityLogStartValue(
+                    type: "Metadata",
+                    request: ActivityLogRequestSnapshot(request: request),
+                    startDatetime: Date(timeIntervalSince1970: 1_700_000_010),
+                    bookId: book.id,
+                    libraryId: library.id
+                )
+            )
+        ])
+
+        let entries = repository.fetchEntries(libraryId: library.id, bookId: book.id, since: .distantPast)
+
+        XCTAssertEqual(entries.first?.libraryName, library.name)
+        XCTAssertEqual(entries.first?.bookTitle, "Mapped Activity Book")
+        XCTAssertEqual(bookRepository.getBookLibraryParam, library)
+        XCTAssertEqual(bookRepository.getBookBookIdParam, book.id)
+    }
+
     func testDeleteAndCleanUseRepositoryBoundary() async throws {
         let oldDate = Date(timeIntervalSince1970: 1_600_000_000)
         let newDate = Date(timeIntervalSince1970: 1_700_000_000)
@@ -108,4 +152,8 @@ final class ActivityLogRepositoryTests: XCTestCase {
         entries = repository.fetchEntries(libraryId: nil, bookId: nil, since: .distantPast)
         XCTAssertTrue(entries.isEmpty)
     }
+}
+
+private final class MockCalibreLibrarySnapshotProvider: CalibreLibrarySnapshotProviding {
+    var calibreLibraries: [String: CalibreLibrary] = [:]
 }
